Assignments/Team Leader - Devanand V/Assignment - 1/Assignment-1-2/app.py
from flask import Flask
import flask_appbuilder 
import flask_bootstrap
import flask_authorize
import flask_login
import flask_security

app = Flask(__name__) 
 
@app.route('/') 
def home():  
    return "Hello world!";  
  
if __name__ =='__main__':  
    app.run(debug = True)  