Assignments/Team Leader - Devanand V/Assignment - 1/Assignment-1-2/requirement.txt
apispec==3.3.2
attrs==22.1.0
Babel==2.10.3
blinker==1.5
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
Flask==2.2.2
Flask-AppBuilder==4.1.4
Flask-Authorize==0.2.6
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.13.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-Security-Too==5.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.26.1
packaging==21.3
passlib==1.7.4
prison==0.2.1
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
